+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/ExoPlayer.java
@@ -40,7 +40,7 @@ import androidx.media3.common.PriorityTaskManager;
 import androidx.media3.common.Timeline;
 import androidx.media3.common.Tracks;
 import androidx.media3.common.VideoSize;
-import androidx.media3.common.text.Cue;
+import androidx.media3.common.text.CueGroup;
 import androidx.media3.common.util.Clock;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
@@ -358,7 +358,7 @@ public interface ExoPlayer extends Player {
      * @deprecated Use {@link Player#getCurrentCues()} instead.
      */
     @Deprecated
-    List<Cue> getCurrentCues();
+    CueGroup getCurrentCues();
   }
 
   /**