+++ libraries/test_session_current/src/main/java/androidx/media3/session/RemoteMediaSession.java
@@ -24,7 +24,7 @@ import static androidx.media3.test.session.common.CommonConstants.KEY_CONTENT_DU
 import static androidx.media3.test.session.common.CommonConstants.KEY_CONTENT_POSITION;
 import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_AD_GROUP_INDEX;
 import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_AD_INDEX_IN_AD_GROUP;
-import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_CUES;
+import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_CUE_GROUP;
 import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_LIVE_OFFSET;
 import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_MEDIA_ITEM_INDEX;
 import static androidx.media3.test.session.common.CommonConstants.KEY_CURRENT_PERIOD_INDEX;
@@ -78,7 +78,7 @@ import androidx.media3.common.Player.PositionInfo;
 import androidx.media3.common.Timeline;
 import androidx.media3.common.TrackSelectionParameters;
 import androidx.media3.common.VideoSize;
-import androidx.media3.common.text.Cue;
+import androidx.media3.common.text.CueGroup;
 import androidx.media3.common.util.BundleableUtil;
 import androidx.media3.common.util.Log;
 import androidx.media3.common.util.UnstableApi;
@@ -389,8 +389,8 @@ public class RemoteMediaSession {
       binder.notifyDeviceVolumeChanged(sessionId, volume, muted);
     }
 
-    public void notifyCuesChanged(List<Cue> cues) throws RemoteException {
-      binder.notifyCuesChanged(sessionId, BundleableUtil.toBundleList(cues));
+    public void notifyCuesChanged(CueGroup cueGroup) throws RemoteException {
+      binder.notifyCuesChanged(sessionId, cueGroup.toBundle());
     }
 
     public void notifyDeviceInfoChanged(DeviceInfo deviceInfo) throws RemoteException {
@@ -595,8 +595,8 @@ public class RemoteMediaSession {
       return this;
     }
 
-    public MockPlayerConfigBuilder setCurrentCues(List<Cue> cues) {
-      bundle.putParcelableArrayList(KEY_CURRENT_CUES, BundleableUtil.toBundleArrayList(cues));
+    public MockPlayerConfigBuilder setCurrentCues(CueGroup cueGroup) {
+      bundle.putBundle(KEY_CURRENT_CUE_GROUP, cueGroup.toBundle());
       return this;
     }