+++ RELEASENOTES.md
@@ -28,6 +28,8 @@
     *   Decrease ad polling rate from every 100ms to every 200ms, to line up
         with Media Rating Council (MRC) recommendations.
 *   Text:
+    *   Change `Player.getCurrentCues()` to return `CueGroup` instead of
+        `List<Cue>`.
     *   SSA: Support `OutlineColour` style setting when `BorderStyle == 3` (i.e.
         `OutlineColour` sets the background of the cue)
         ([#8435](https://github.com/google/ExoPlayer/issues/8435)).